+++ src/readline/rlprivate.h
@@ -170,7 +170,7 @@ extern void _rl_output_character_function PARAMS((int));
 #else
 extern int _rl_output_character_function PARAMS((int));
 #endif
-extern void _rl_output_some_chars PARAMS((const char *, int));
+extern void _rl_output_some_chars PARAMS((const char *, size_t));
 extern int _rl_backspace PARAMS((int));
 extern void _rl_enable_meta_key PARAMS((void));
 extern void _rl_control_keypad PARAMS((int));
